only download sami if it doesn’t exist

+++ build/bin/build-docs.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
 
 # Use the Phar rather than require it because we don't deploy the docs on every build.
-curl -O http://get.sensiolabs.org/sami.phar
+if [ ! -f sami.phar ]; then
+    curl -O http://get.sensiolabs.org/sami.phar
+fi
 
 # Because Travis only clones a single branch, we only want to build everything on master.
 if [ "$TRAVIS_BRANCH" == "master" ]; then
